Turn functionality changed
Battle system is being redesigned so that the player selects actions for all characters, and the turn plays out.

Currently, the game spawns in characters on start, then awaits the player pressing spacebar to play out one single turn of combat, with each character acting once. BeginTurns function has been recreated and functionality of PlayerTurn and EnemyTurn readjusted to suit this. ActionSelect state has been created to await player input.

To do next: Create functionality for player actually selecting specific abilities, adding those abilities to a array for each character which is then played out when the turn begins and the character acts.

Some of the blueprints are a bit messy as well currently, due to leaving old, unhooked nodes in them in case they need to be reused later. They should be removed in the next push.