Turn System Rework
Battle System has been reworked so that HP is team based, not for each individual unit. Damage is now calculated by calculating which team's attacks did higher damage that turn, subtracting the other team's damage from that, and that is how much damage is dealt to the team whose damage values were lower.
ActionSelect Event created in the event graph for re-enabling action select UI and adding new actions to what is available to the player in the future.
NextTurn Event now calls the reworked PlayerTurn, EnemyTurn and then the new function CalculateDamage, one by one and then checks if the battle is over.