Functionality for Enemy Action Select in progress. HP variables updated
Enemy now draws a random action from their deck for an attack. (There's a big chunk of unused nodes currently in the EnemyTurn function. Ignore it for now, its intended to have the enemy select different amount of actions but it isnt functional yet).
Player and Enemy HP variables renamed to CurrPlayersHP and CurrEnemiesHP, and PlayersMaxHP and EnemiesMaxHP variables added. All 4 variables default to a value of 50