Card Sort testing
Card sort functionality has now been properly tested. CardDamageCalc has been added as a component of BattlePlayerController and the SortCards function is called on the player turn (this latter bit was for testing and will be removed soon.) Next is to properly do functionality for checking for straights and then fully implementing carddamagecalc into the battleplayercontroller and doing some refactoring.